backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async and tasks
redis==5.0.1
celery==5.3.6

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code quality
black==23.12.1
mypy==1.8.0
ruff==0.1.11
